+++ today's horoscope/apps/react-web/src/pages/detail/components/textbox/star/StarText.tsx
@@ -1,7 +1,9 @@
 import React from 'react';
+
 import { useQuery } from '@tanstack/react-query';
 import QUERY_KEYS from '../../../../../services/queryKeys';
 import APIS from '../../../../../services/api';
+
 import Styles from './StarText.module.scss';
 
 interface StarFortunes {
