+++ today's horoscope/apps/react-web/src/pages/detail/main/zodiac/Zodiac.module.scss
@@ -6,7 +6,7 @@
   width: 100%;
   height: 194px;
   position: fixed;
-  background: linear-gradient(180deg, #dbe1ff 0%, #eff2ff 100%);
+  background: linear-gradient(180deg, #ffe5f1 0%, #ffe0f4 100%);
 }
 
 .headicon {
@@ -15,7 +15,7 @@
   position: fixed;
   top: 30px;
   left: 0;
-  right: 0; /* 화면 전체 너비를 사용 */
+  right: 0;
   padding: 0;
   justify-content: space-between;
 }
@@ -63,7 +63,7 @@
   flex-direction: column;
   margin-left: auto;
   margin-right: auto;
-  max-width: calc(100% - 64px); /* 양쪽 끝 넓이를 32px로 유지 */
+  max-width: calc(100% - 64px);
   display: inline-flex;
   justify-content: space-around; /* 혹은 space-between */
   align-items: center;
@@ -75,9 +75,9 @@
   height: 80px;
   display: inline-flex;
   align-items: center;
-  margin-top: 35.5px; /* 상단 간격을 20px로 조정 */
-  margin-left: 20px; /* 좌측 간격 유지 */
-  margin-right: 20px; /* 우측 간격 유지 */
+  margin-top: 35.5px;
+  margin-left: 20px;
+  margin-right: 20px;
 }
 
 .zodiacText {
